requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
python-dotenv==1.0.1
python-multipart==0.0.12
qrcode[pil]==8.0
pillow==11.0.0
PyJWT==2.10.1
jinja2==3.1.4
